use buffer in parsing instead of strings
